requirements.txt
beautifulsoup4
matplotlib
numpy
pandas
pycparser
lmfit
sticks
statsmodels
jupyter
tqdm
python-dateutil
anytree
seaborn